BSC mainnet script
